+++ packages/node/src/rest/types/auth/index.ts
@@ -5,5 +5,5 @@ import { User } from "../users";
 export type AuthEndpoints =
   | Endpoint<"POST", "/auth/sign-up", User, CreateUserDto>
   | Endpoint<"POST", "/auth/sign-in", User, SignInUserDto>
-  | Endpoint<"POST", "/auth/sign-out", void>
-  | Endpoint<"POST", "/auth/refresh-token", User>;
+  | Endpoint<"POST", "/auth/sign-out", null, null>
+  | Endpoint<"POST", "/auth/refresh-token", null, null>;