+++ packages/node/CHANGELOG.md
@@ -1,5 +1,59 @@
 # tonightpass
 
+## 0.0.16
+
+### Patch Changes
+
+- [`bc0b252`](https://github.com/tonightpass/tonightpass/commit/bc0b2527533944a433abc10e69a8dceeec773da2) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Update enums names
+
+## 0.0.15
+
+### Patch Changes
+
+- [#232](https://github.com/tonightpass/tonightpass/pull/232) [`7c2353a`](https://github.com/tonightpass/tonightpass/commit/7c2353a308119eef0cb5515ad0140e8e8fd1093b) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Add organization members sdk
+
+## 0.0.14
+
+### Patch Changes
+
+- [`79c52a5`](https://github.com/tonightpass/tonightpass/commit/79c52a50ec141a6e817b29377d3facde337c5c9e) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Add decorators support
+
+## 0.0.13
+
+### Patch Changes
+
+- [`858e5f2`](https://github.com/tonightpass/tonightpass/commit/858e5f20f89db91a6d510453d8d6c089aa2792ce) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Update profile type description to be optional
+
+## 0.0.12
+
+### Patch Changes
+
+- [`2c5b552`](https://github.com/tonightpass/tonightpass/commit/2c5b55231112404a705a1b822c67fdcd72cd467c) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Update profiles types with Organization identity
+
+## 0.0.11
+
+### Patch Changes
+
+- [`e05e50e`](https://github.com/tonightpass/tonightpass/commit/e05e50eb071f55b46686f14f25ba01c1402cf84b) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Fix optional fields in CreateOrganizationDto
+
+## 0.0.10
+
+### Patch Changes
+
+- [#218](https://github.com/tonightpass/tonightpass/pull/218) [`d4d27fb`](https://github.com/tonightpass/tonightpass/commit/d4d27fb683212d88bf4c4650ce33ab376c23f445) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Fix Organization types and dtos
+
+## 0.0.9
+
+### Patch Changes
+
+- [`8f8bd1b`](https://github.com/tonightpass/tonightpass/commit/8f8bd1b21c684389c44c7b9ef4be3452061260e1) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Add organizations sdk
+
+## 0.0.8
+
+### Patch Changes
+
+- [`d009e78`](https://github.com/tonightpass/tonightpass/commit/d009e7808d5874915dc7a40e244330a8c164cdff) Thanks [@AntoineKM](https://github.com/AntoineKM)! - Fix empty auth post requests
+
 ## 0.0.7
 
 ### Patch Changes